Change names for classes being refactored (_(NAME) are being deprecated)
