+++ DNM_iOS/DNM_iOS/ButtonSwitchNodeComplex.swift
@@ -30,7 +30,7 @@ public class ButtonSwitchNodeComplex: UIView {
     private let buttonSwitchNode_width: CGFloat = 75
     
     // test
-    public var systemView: _SystemView?
+    public var systemView: SystemUIView?
     public var componentTypesShownByID: [String : [String]] {
         get {
             var componentTypesShownByID: [String: [String]] = [:]
