+++ DNM_iOS/DNM_iOS/RadioGroupPanel.swift
@@ -57,7 +57,7 @@ public class RadioGroupPanel: ButtonSwitchPanel {
             statesByText[sender.text] = sender.isOn
         }
     
-        if let environment = target as? Environment {
+        if let environment = target as? _Environment {
             environment.goToViewWithID(currentButtonSelectedID)
         }
     }
